fix: change .catch in controllers

+++ controllers/cards.js
@@ -1,10 +1,13 @@
-const Cards = require('../models/card');
-const BadRequestError = require('../errors/bad-request-error');
-const ForbiddenError = require('../errors/forbidden-error');
-const NotFoundError = require('../errors/not-found-error');
+const Card = require('../models/card');
+
+const {
+  BadRequestError,
+  ForbiddenError,
+  NotFoundError,
+} = require('../errors/index-errors');
 
 module.exports.getAllCards = (req, res, next) => {
-  Cards
+  Card
     .find({})
     .then((cards) => res.status(200).send(cards))
     .catch(next);
@@ -14,76 +17,74 @@ module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
   const owner = req.user._id;
 
-  Cards
+  Card
     .create({ name, link, owner })
     .then((card) => res.status(200).send(card))
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        throw new BadRequestError(`Переданы некорректные данные при создании карточки -- ${err.name}`);
-      } else {
-        return next(err);
+        next(new BadRequestError(`Переданы некорректные данные при создании карточки -- ${err.name}`));
       }
+      next(err);
     });
 };
 
 module.exports.deleteCardById = (req, res, next) => {
   const { cardId } = req.params;
 
-  Cards
+  Card
     .findById(cardId)
-    .orFail(() => {
-      throw new NotFoundError('Карточка с указанным _id не найдена');
-    })
     .then((card) => {
+      if (!card) {
+        return next(new NotFoundError('Карточка с указанным _id не найдена'));
+      }
       if (card.owner.toString() === req.user._id) {
-        Cards
+        Card
           .findByIdAndRemove(cardId)
           .then(() => res.status(200).send(card));
-      } else {
-        throw new ForbiddenError('Попытка удалить чужую карточку');
       }
+      throw new ForbiddenError('Попытка удалить чужую карточку');
     })
     .catch(next);
 };
 
 module.exports.likeCard = (req, res, next) => {
-  Cards
+  Card
     .findByIdAndUpdate(
       req.params.cardId,
       { $addToSet: { likes: req.user._id } },
       { new: true },
     )
-    .orFail(() => {
-      throw new NotFoundError('Карточка с указанным _id не найдена');
+    .then((card) => {
+      if (!card) {
+        next(new NotFoundError('Карточка с указанным _id не найдена'));
+      }
+      res.status(200).send(card);
     })
-    .then((card) => res.status(200).send(card))
     .catch((err) => {
       if (err.name === 'CastError') {
         next(new BadRequestError('Переданы некорректные данные для постановки лайка'));
-      } else {
-        next(err);
       }
+      next(err);
     });
 };
 
 module.exports.dislikeCard = (req, res, next) => {
-  Cards
+  Card
     .findByIdAndUpdate(
       req.params.cardId,
       { $pull: { likes: req.user._id } },
       { new: true },
     )
-    .orFail(() => {
-      throw new NotFoundError('Передан несуществующий _id карточки');
+    .then((card) => {
+      if (!card) {
+        next(new NotFoundError('Карточка с указанным _id не найдена'));
+      }
+      res.status(200).send(card);
     })
-    .then((card) => res.status(200).send(card))
     .catch((err) => {
       if (err.name === 'CastError') {
         next(new BadRequestError('Переданы некорректные данные для снятия лайка'));
-      // } else if (err.message === 'NotFound') {
-      //   next(new NotFoundError('Передан несуществующий _id карточки'));
-      } else {
-        next(err);
       }
+      next(err);
     });
 };
